refactor: move source code to own module directory
