+++ cli_test.bat
@@ -1,3 +1,3 @@
 ﻿@echo off & chcp 65001 >nul
 
-python noScribe.py "C:\Users\kai\Documents\Programmierung\2023_WhisperTranscribe\Intw\Ein Gespräch mit Heikedine Körting kurz.mp3" "C:\Users\kai\Documents\Programmierung\2023_WhisperTranscribe\Intw\t.html"
+python -m noScribe "C:\Users\kai\Documents\Programmierung\2023_WhisperTranscribe\Intw\Ein Gespräch mit Heikedine Körting kurz.mp3" "C:\Users\kai\Documents\Programmierung\2023_WhisperTranscribe\Intw\t.html"
