+++ noScribe/__main__.py
@@ -1,0 +1,4 @@
+from . import noScribe
+
+if __name__ == "__main__":
+    noScribe.noScribeMain()
