+++ noScribe/pyannote_mp_worker.py
