+++ noScribe/tkHyperlinkManager.py
