+++ noScribe/utils.py
