+++ noScribe/whisper_mp_worker.py
