+++ pyinstaller/noScribe_linux.spec
@@ -14,7 +14,7 @@ datas += tmp_ret[0]; binaries += tmp_ret[1]; hiddenimports += tmp_ret[2]
 
 
 a = Analysis(
-    ['../noScribe.py'],
+    ['../noScribe/noScribe.py'],
     pathex=['..'],
     binaries=binaries,
     datas=datas,
